Break out DTO's into separate project, make controller action simple and asynchronous.
