+++ Angular.Demo.API/Angular.Demo.API.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Angular.Demo.DataContract\Angular.Demo.DataContract.csproj" />
+  </ItemGroup>
+
 </Project>
