+++ Angular.Demo.API/Controllers/WeatherForecastController.cs
@@ -1,3 +1,4 @@
+using Angular.Demo.DataContract;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Angular.Demo.API.Controllers
@@ -5,10 +6,6 @@ namespace Angular.Demo.API.Controllers
 
 	public class WeatherForecastController : ApiControllerBase
 	{
-		private static readonly string[] Summaries = new[]
-		{
-				"Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-		};
 
 		private readonly ILogger<WeatherForecastController> _logger;
 		private readonly AppSettingsConfig _appSettings;
@@ -20,15 +17,10 @@ namespace Angular.Demo.API.Controllers
 		}
 
 		[HttpGet]
-		public IEnumerable<WeatherForecast> Get()
+		public async Task<ActionResult<IEnumerable<WeatherForecast>>> Get()
 		{
-			return Enumerable.Range(1, 5).Select(index => new WeatherForecast
-			{
-				Date = DateTime.Now.AddDays(index),
-				TemperatureC = Random.Shared.Next(-20, 55),
-				Summary = Summaries[Random.Shared.Next(Summaries.Length)]
-			})
-			.ToArray();
+			return Ok(await WeatherForecast.GetWeatherForecastsAsync(5));
 		}
+
 	}
 }
