+++ AngularDemo.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 17.1.32328.378
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{54A90642-561A-4BB1-A94E-469ADEE60C69}") = "Angular.Demo.UI", "Angular.Demo.UI\Angular.Demo.UI.esproj", "{9C80B1C2-E761-4DE6-8423-922346FCD298}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Angular.Demo.API", "Angular.Demo.API\Angular.Demo.API.csproj", "{3D095A2E-C096-4208-8777-411B2B61B938}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Angular.Demo.API", "Angular.Demo.API\Angular.Demo.API.csproj", "{3D095A2E-C096-4208-8777-411B2B61B938}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Angular.Demo.DataContract", "Angular.Demo.DataContract\Angular.Demo.DataContract.csproj", "{9EA13EC6-14B6-4B40-8E2E-CDFB62345631}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,6 +25,10 @@ Global
 		{3D095A2E-C096-4208-8777-411B2B61B938}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3D095A2E-C096-4208-8777-411B2B61B938}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3D095A2E-C096-4208-8777-411B2B61B938}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9EA13EC6-14B6-4B40-8E2E-CDFB62345631}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9EA13EC6-14B6-4B40-8E2E-CDFB62345631}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9EA13EC6-14B6-4B40-8E2E-CDFB62345631}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9EA13EC6-14B6-4B40-8E2E-CDFB62345631}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
